requirements.txt
flask
openai
pinecone-client
tqdm
